+++ bin/localization_gen.dart
@@ -1,0 +1,8 @@
+#!/usr/bin/env dart
+
+import 'package:localization_gen/src/command/generate_command.dart';
+
+void main(List<String> args) {
+  GenerateCommand().run(args);
+}
+